search on the navbar completed
