+++ src/App.module.scss
@@ -14,7 +14,7 @@ body {
 .App {
   display: grid;
   grid-template-columns: repeat(12, 1fr);
-  grid-template-rows: 70px auto 70px;
+  grid-template-rows: auto 70px;
   font-weight: 600;
 }
 
@@ -22,15 +22,15 @@ header {
   background-color: $backgroundNavbar;
   width: 100%;
   text-align: center;
-  position: fixed;
-  z-index: 999;
+  position: absolute;
 }
 
 main {
   grid-column: 2 / 12;
   background-color: #4d4d4d;
-  margin: 5px 0;
-  min-height: calc(100vh - 150px);
+  margin-top: 100px;
+  margin-bottom: 10px;
+  min-height: calc(100vh - 70px);
 }
 
 footer {
