+++ src/components/SearchShow.tsx
@@ -1,23 +1,12 @@
-import React, { useState } from "react";
-import axios from "axios";
+import React from "react";
 import _ from "lodash";
-import ReactHtmlParser from "react-html-parser"
 import { MovieInterface, ResponseInterface } from "../types/interfaces";
-import styles from "../styles/SearchShow.module.css";
 import MovieCard from "./MovieCard";
 import { Context, AppActionInterface } from "./../Store";
-import Favorites from './FavoritesSidebar';
 
-type FormElem = React.FormEvent<HTMLFormElement>;
-type ChangeEvent = React.ChangeEvent<HTMLInputElement>;
 
 const SearchShow = (): JSX.Element => {
-  const [searchInput, setSearchInput] = useState<string>("");
   const { state, dispatch } = React.useContext(Context);
-
-  const handleChange = (e: ChangeEvent): void => {
-    setSearchInput(e.target.value);
-  };
 
   const toggleFavorite = (show: MovieInterface): AppActionInterface => {
     if (state.favorites.includes(show)) {
@@ -33,43 +22,12 @@ const SearchShow = (): JSX.Element => {
     }
   };
 
-  const handleSearchMovies = async (e: FormElem) => {
-    e.preventDefault();
-
-    const url = `http://api.tvmaze.com/search/shows?q=${searchInput}`;
-    const response = await axios.get(url);
-    const results = response.data;
-    const movies = results.map((result: ResponseInterface) => result.show);
-    setSearchInput("");
-
-    if(_.isEmpty(movies)) {
-      console.log("<h3>No Movies Found...</h3>")
-    }
-
-    return dispatch({
-      type: "SEARCH_MOVIE",
-      payload: movies
-    });
-  };
-
   return (
-    <div className={styles["form-container"]}>
-      <form onSubmit={handleSearchMovies} className={styles["search-form"]}>
-        <input
-          type="text"
-          onChange={handleChange}
-          value={searchInput}
-          placeholder="Search tv shows.."
-          required
-        />
-        <button type="submit">Search</button>
-      </form>
       <MovieCard
         searchResult={state.movies}
         favorites={state.favorites}
         toggleFavorite={toggleFavorite}
       />
-    </div>
   );
 };
