+++ src/styles/SearchForm.module.scss
@@ -4,7 +4,7 @@ form {
   display: flex;
   flex-flow: row nowrap;
   justify-content: flex-end;
-  width: calc(100% - 20px);
+  width: 40%;
   border-radius: 5px;
   background-color: #fff;
 }
@@ -24,11 +24,25 @@ button {
   flex: 1 1 10%;
   height: 35px;
   margin: 7px 7px 0 0;
-  background-color: $searchButtonColor;
+  background-color: $backgroundYellow;
   border-radius: 3px;
   box-sizing: border-box;
   font-size: 16px;
   font-weight: 600;
   cursor: pointer;
   border: none;
+}
+
+@media screen and (min-width: 768) {
+  button {
+    flex: 1 1 3%;
+  }
+  form {
+    display: flex;
+    flex-flow: row nowrap;
+    justify-content: flex-end;
+    width: calc(100% - 20px);
+    border-radius: 5px;
+    background-color: #fff;
+  }
 }
