+++ src/styles/_variables.scss
@@ -1,7 +1,6 @@
 $backgroundGrey: rgb(240, 240, 240);
 $backgroundDarkGrey: rgb(51, 51, 51);
 $backgroundGreen: rgb(14, 175, 22);
-$backgroundYellow: rgb(250, 236, 45);
+$backgroundYellow: rgb(240, 188, 17);
 
-$backgroundNavbar: #0bbed6;
-$searchButtonColor: rgb(240, 106, 17);
+$backgroundNavbar: #000000;
